requirements/base.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    make upgrade
#
analytics-python==1.2.9
    # via -r requirements/base.in
apache-bookkeeper-client==4.12.1
    # via pulsar-client
attrs==20.3.0
    # via edx-ace
bleach==3.3.0
    # via -r requirements/base.in
certifi==2020.12.5
    # via
    #   pulsar-client
    #   requests
cffi==1.14.4
    # via cryptography
chardet==3.0.4
    # via requests
coreapi==2.3.3
    # via
    #   -r requirements/base.in
    #   django-rest-swagger
    #   openapi-codec
coreschema==0.0.4
    # via coreapi
cryptography==3.4.4
    # via
    #   pyjwt
    #   social-auth-core
defusedxml==0.6.0
    # via
    #   python3-openid
    #   social-auth-core
django-appconf==1.0.4
    # via django-statici18n
django-compat==1.0.15
    # via django-hijack
django-crum==0.7.9
    # via edx-django-utils
django-debug-toolbar==3.2
    # via -r requirements/base.in
django-extensions==3.1.1
    # via -r requirements/base.in
django-filter==2.4.0
    # via -r requirements/base.in
django-hijack==2.2.1
    # via -r requirements/base.in
git+https://github.com/jsocol/django-ratelimit@main
    # via -r requirements/base.in
django-rest-swagger==2.2.0
    # via -r requirements/base.in
django-simple-history==2.12.0
    # via -r requirements/base.in
django-sortedm2m==3.0.2
    # via -r requirements/base.in
django-statici18n==2.0.1
    # via -r requirements/base.in
django-storages==1.11.1
    # via -r requirements/base.in
django-waffle==2.1.0
    # via
    #   -r requirements/base.in
    #   edx-django-utils
    #   edx-drf-extensions
django-webpack-loader==0.7.0
    # via -r requirements/base.in
django==2.2.18
    # via
    #   -c https://raw.githubusercontent.com/edx/edx-lint/master/edx_lint/files/common_constraints.txt
    #   -r requirements/base.in
    #   django-appconf
    #   django-crum
    #   django-debug-toolbar
    #   django-filter
    #   django-hijack
    #   django-statici18n
    #   django-storages
    #   djangorestframework
    #   drf-jwt
    #   edx-ace
    #   edx-auth-backends
    #   edx-credentials-themes
    #   edx-django-release-util
    #   edx-django-sites-extensions
    #   edx-django-utils
    #   edx-drf-extensions
    #   edx-i18n-tools
    #   rest-condition
    #   xss-utils
djangorestframework==3.12.2
    # via
    #   -r requirements/base.in
    #   django-rest-swagger
    #   drf-jwt
    #   edx-drf-extensions
    #   rest-condition
drf-jwt==1.17.3
    # via edx-drf-extensions
edx-ace==0.1.17
    # via -r requirements/base.in
edx-auth-backends==3.3.0
    # via
    #   -c requirements/constraints.txt
    #   -r requirements/base.in
edx-django-release-util==1.0.0
    # via -r requirements/base.in
edx-django-sites-extensions==3.0.0
    # via -r requirements/base.in
edx-django-utils==3.13.0
    # via
    #   -r requirements/base.in
    #   edx-drf-extensions
    #   edx-rest-api-client
edx-drf-extensions==6.4.0
    # via -r requirements/base.in
edx-i18n-tools==0.5.3
    # via edx-credentials-themes
edx-opaque-keys==2.2.0
    # via
    #   -r requirements/base.in
    #   edx-drf-extensions
edx-rest-api-client==5.3.0
    # via -r requirements/base.in
git+https://github.com/edx/credentials-themes.git@0.1.52#egg=edx_credentials_themes==0.1.52
    # via -r requirements/base.in
fastavro==0.24.0
    # via pulsar-client
future==0.18.2
    # via pyjwkest
gitdb==4.0.5
    # via gitpython
gitpython==3.1.13
    # via transifex-client
grpcio==1.27.2
    # via
    #   apache-bookkeeper-client
    #   pulsar-client
idna==2.10
    # via requests
itypes==1.2.0
    # via coreapi
jinja2==2.11.3
    # via coreschema
markdown==3.3.3
    # via -r requirements/base.in
markupsafe==1.1.1
    # via jinja2
mysqlclient==2.0.3
    # via -r requirements/base.in
newrelic==6.0.1.155
    # via
    #   -r requirements/base.in
    #   edx-django-utils
oauthlib==3.1.0
    # via
    #   requests-oauthlib
    #   social-auth-core
openapi-codec==1.3.2
    # via django-rest-swagger
packaging==20.9
    # via bleach
path.py==12.5.0
    # via edx-i18n-tools
path==15.1.0
    # via path.py
pbr==5.5.1
    # via stevedore
pillow==8.1.0
    # via -r requirements/base.in
polib==1.1.0
    # via edx-i18n-tools
prometheus-client==0.9.0
    # via pulsar-client
protobuf==3.14.0
    # via
    #   apache-bookkeeper-client
    #   pulsar-client
psutil==5.8.0
    # via edx-django-utils
pulsar-client==2.7.0
    # via -r requirements/base.in
pycparser==2.20
    # via cffi
pycryptodomex==3.10.1
    # via pyjwkest
pygments==2.7.4
    # via -r requirements/base.in
pyjwkest==1.4.2
    # via edx-drf-extensions
pyjwt[crypto]==1.7.1
    # via
    #   drf-jwt
    #   edx-auth-backends
    #   edx-rest-api-client
    #   social-auth-core
pymmh3==0.0.5
    # via apache-bookkeeper-client
pymongo==3.11.3
    # via edx-opaque-keys
pyparsing==2.4.7
    # via packaging
python-dateutil==2.8.1
    # via
    #   analytics-python
    #   edx-ace
    #   edx-drf-extensions
python-memcached==1.59
    # via -r requirements/base.in
python-slugify==4.0.1
    # via transifex-client
python3-openid==3.2.0
    # via social-auth-core
pytz==2021.1
    # via
    #   -r requirements/base.in
    #   apache-bookkeeper-client
    #   django
    #   fastavro
pyyaml==5.4.1
    # via
    #   edx-django-release-util
    #   edx-i18n-tools
ratelimit==2.2.1
    # via pulsar-client
requests-oauthlib==1.3.0
    # via social-auth-core
requests==2.24.0
    # via
    #   -c requirements/constraints.txt
    #   -r requirements/base.in
    #   analytics-python
    #   apache-bookkeeper-client
    #   coreapi
    #   edx-drf-extensions
    #   edx-rest-api-client
    #   pyjwkest
    #   requests-oauthlib
    #   sailthru-client
    #   slumber
    #   social-auth-core
    #   transifex-client
rest-condition==1.0.3
    # via edx-drf-extensions
sailthru-client==2.2.3
    # via edx-ace
semantic-version==2.8.5
    # via edx-drf-extensions
simplejson==3.17.2
    # via
    #   django-rest-swagger
    #   sailthru-client
six==1.15.0
    # via
    #   analytics-python
    #   apache-bookkeeper-client
    #   bleach
    #   django-compat
    #   django-simple-history
    #   edx-ace
    #   edx-auth-backends
    #   edx-django-release-util
    #   edx-drf-extensions
    #   edx-i18n-tools
    #   grpcio
    #   protobuf
    #   pulsar-client
    #   pyjwkest
    #   python-dateutil
    #   python-memcached
    #   social-auth-app-django
    #   social-auth-core
    #   transifex-client
slumber==0.7.1
    # via edx-rest-api-client
smmap==3.0.5
    # via gitdb
social-auth-app-django==4.0.0
    # via
    #   -r requirements/base.in
    #   edx-auth-backends
social-auth-core==4.0.2
    # via
    #   -c https://raw.githubusercontent.com/edx/edx-lint/master/edx_lint/files/common_constraints.txt
    #   -c requirements/constraints.txt
    #   edx-auth-backends
    #   social-auth-app-django
sqlparse==0.4.1
    # via
    #   django
    #   django-debug-toolbar
stevedore==3.3.0
    # via
    #   edx-ace
    #   edx-django-utils
    #   edx-opaque-keys
text-unidecode==1.3
    # via python-slugify
transifex-client==0.14.2
    # via edx-credentials-themes
uritemplate==3.0.1
    # via coreapi
urllib3==1.25.11
    # via
    #   requests
    #   transifex-client
webencodings==0.5.1
    # via bleach
xss-utils==0.2.0
    # via -r requirements/base.in

# The following packages are considered to be unsafe in a requirements file:
# setuptools
